+++ search.php
@@ -27,7 +27,7 @@ get_header();
 				<?php
 				while ( have_posts() ) :
 					the_post();
-					get_template_part( 'template-parts/content/content', 'search' );
+					get_template_part( 'template-parts/content/content', 'row' );
 				endwhile;
 				?>
 			</div>